Better output, vacuum database, no session with no errors

+++ FA_db/repair.py
@@ -45,39 +45,45 @@ def find_errors(DB):
     return errs_id, errs_vl, errs_fl
 
 def dberrors(DB):
-    print()
+    print('Analyzing database for errors ... ', end='', flush=True)
     errs_id, errs_vl, errs_fl = find_errors(DB)
+    print('Done')
 
     print(f'There are {len(errs_id)} id errors')
     print(f'There are {len(errs_vl)} field value errors')
     print(f'There are {len(errs_fl)} files errors')
 
-    print()
-    Session = session()
-    print()
+    if len(errs_id) or len(errs_vl) or len(errs_fl):
+        print()
+        Session = session()
+        print()
 
-    if not Session:
-        print('Session error')
-        return
+        if not Session:
+            print('Session error')
+            return
 
-    if len(errs_vl):
-        print('Fixing field values errors')
-        i, l = 0, len(str(len(errs_vl)))
-        for sub in errs_vl:
-            sub = str(sub[0])
-            i += 1
-            print(f'{i:0>{l}}/{len(errs_vl)} - {sub:0>10}\r', end='', flush=True)
-            if not check_page(Session, 'view/'+sub): continue
-            DB.execute(f'DELETE FROM submissions WHERE id = {sub[0]}')
-            DB.commit()
-            dl_sub(Session, sub, f'FA.files/{tiers(sub[0])}/{sub[0]:0>10}', DB, True, False, 2)
-        print(' '*(l+l+1+3+10), end='\r', flush=True)
+        if len(errs_vl):
+            print('Fixing field values errors')
+            i, l = 0, len(str(len(errs_vl)))
+            for sub in errs_vl:
+                sub = str(sub[0])
+                i += 1
+                print(f'{i:0>{l}}/{len(errs_vl)} - {sub:0>10}\r', end='', flush=True)
+                if not check_page(Session, 'view/'+sub): continue
+                DB.execute(f'DELETE FROM submissions WHERE id = {sub[0]}')
+                DB.commit()
+                dl_sub(Session, sub, f'FA.files/{tiers(sub[0])}/{sub[0]:0>10}', DB, True, False, 2)
+            print(' '*(l+l+1+3+10), end='\r', flush=True)
 
-    # if len(errs_fl):
-    #     print('Fixing missing files')
-    #     i, l = 0, len(str(len(errs_fl)))
-    #     for sub in errs_fl:
-    #         i += 1
-    #         print(f'{i:0>{l}}/{len(errs_fl)} - {sub:0>10}\r', end='', flush=True)
-    #         dl_sub(Session, str(sub), f'FA.files/{tiers(sub[0])}/{sub[0]:0>10}', DB, True, False, 2)
-    #     print(' '*(l+l+1+3+10), end='\r', flush=True)
+        # if len(errs_fl):
+        #     print('Fixing missing files')
+        #     i, l = 0, len(str(len(errs_fl)))
+        #     for sub in errs_fl:
+        #         i += 1
+        #         print(f'{i:0>{l}}/{len(errs_fl)} - {sub:0>10}\r', end='', flush=True)
+        #         dl_sub(Session, str(sub), f'FA.files/{tiers(sub[0])}/{sub[0]:0>10}', DB, True, False, 2)
+        #     print(' '*(l+l+1+3+10), end='\r', flush=True)
+
+    print('Optimizing database ... ', end='', flush=True)
+    DB.execute("VACUUM") ; print('Done')
+    print('All done')
